Change the way the Oracle-specific version of Java is installed
Oracle now requires you to create an account in order to download the
Java archives that they distribute.  I created such an account and
downloaded the archive here so that it can be copied over and
installed.

Also update the test code, since the two packages that are installed
have changed names slightly.  (The text "-local" has been appended to
their names.)  As a result the tests need to be altered slightly to
agree with this.